Rendered SearchBar functionality real good
